examples better setup
