+++ .gitignore
@@ -5,5 +5,6 @@ models
 build
 dist
 *.egg-info
+*.png
 __pycache__
 add_license.py
